fix(api): fire BrowserContext.Page event in WebKit and Firefox
